pyproject.toml
[project]
name = "energy-behavior-prosumers"
version = "0.1.0"
description = "Energy behavior prediction for prosumers using machine learning"
authors = [
    {name = "Energy ML Team"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "seaborn>=0.12.0",
    "matplotlib>=3.6.0",
    "colorama>=0.4.6",
    "xgboost>=1.7.0,<3.0.0",
    "lightgbm>=4.0.0",
    "torch>=2.0.0",
    "geopy>=2.3.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "kaggle>=1.7.4.5",
    "mlflow>=2.8.0",
    "pymysql>=1.1.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 
